requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
schedule==1.2.1
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
tenacity==8.2.3 